+++ portfolio_analyzer/portfolio_analyzer/metrics.py
@@ -4,7 +4,7 @@ Provides CAGR (Compound Annual Growth Rate) and XIRR (Extended Internal Rate of 
 calculation functions.
 
 Author: Zhuo Robert Li
-Version: 1.3.3
+Version: 1.3.4
 """
 
 import numpy as np