+++ portfolio_analyzer/tests/__init__.py
@@ -15,7 +15,7 @@ Run specific module:
     python3 -m unittest tests.test_utils -v
 
 Author: Zhuo Robert Li
-Version: 1.2.0
+Version: 1.3.4
 License: ISC
 """