+++ portfolio_analyzer/tests/test_utils.py
@@ -5,7 +5,7 @@ Run with:
     python3 -m unittest test_utils.py -v
 
 Author: Zhuo Robert Li
-Version: 1.2.0
+Version: 1.3.4
 License: ISC
 """